refactor: evict main.rs contents
